+++ src/compiler/parser_internal.h
@@ -79,11 +79,12 @@ INLINE void add_decl_to_list(Decl ***list, Decl *decl)
 }
 
 bool parse_module(ParseContext *c, AstId contracts);
-Expr *parse_generic_parameter(ParseContext *c);
+
 bool try_consume(ParseContext *c, TokenType type);
 bool consume(ParseContext *c, TokenType type, const char *message, ...);
 bool consume_const_name(ParseContext *c, const char* type);
 Expr *parse_precedence_with_left_side(ParseContext *c, Expr *left_side, Precedence precedence);
+Expr *parse_generic_parameter(ParseContext *c);
 
 INLINE const char *symstr(ParseContext *c)
 {