+++ releasenotes.md
@@ -3,6 +3,8 @@
 ## 0.5.0 Change List
 
 ### Changes / improvements
+- New generic syntax.
+- Ad hoc generics are now allowed.
 - Allow inferred type on method first argument.
 - Fix to void expression blocks
 - Temporary objects may now invoke methods using ref parameters.